requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
dj-email-url==1.0.2
Django==3.2.9
django-allauth==0.45.0
django-crispy-forms==1.13.0
django-versatileimagefield==2.2
django-widget-tweaks==1.4.9
idna==3.3
oauthlib==3.2.0
Pillow==8.4.0
platformdirs==2.4.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.21
PyJWT==2.3.0
python-magic==0.4.25
python3-openid==3.2.0
pytz==2021.3
regex==2021.10.23
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.2
tomli==1.2.2
typing-extensions==3.10.0.2
urllib3==1.26.7
whitenoise==5.3.0
